Reorganize codebase into layered architecture (api/, download/, cli/)
Complete Phase 5 of the refactor - file reorganization:

**Directory Structure**:
- src/api/ - API client and validation layer
  - client.js (OpenProcessingClient - 14 endpoints)
  - validator.js (centralized validation)
  - types.js (JSDoc type definitions)

- src/download/ - Download orchestration and content generation
  - service.js (DownloadService - multi-call orchestration)
  - downloader.js (main download logic)
  - htmlGenerator.js, codeAttributor.js, licenseHandler.js
  - metadataWriter.js, serverRunner.js, viteScaffolder.js

- src/cli/ - CLI presentation layer
  - index.js (CLI parser)
  - formatters.js, fieldRegistry.js, fieldSelector.js
  - commands/ (sketch.js, user.js, curation.js, fields.js)

- src/ root - Public API and utilities
  - index.js (main opdl function export)
  - utils.js (shared utilities)
  - fetcher.js (deprecated, for backward compatibility)

**Updated All Imports**:
- Fixed 30+ import paths across source files
- Updated bin/opdl.js to reference new paths
- Migrated 17 test files to mirror source structure
- All 368 tests passing after reorganization

**Benefits**:
- Clear separation of concerns (API → Service → CLI)
- Easier navigation and maintenance
- Test structure mirrors source structure
- Prepared for future enhancements
- Maintains backward compatibility
