+++ src/cli/commands/fields.js
