+++ src/cli/fieldRegistry.js
