+++ src/cli/fieldSelector.js
