+++ src/download/licenseHandler.js
