+++ src/download/metadataWriter.js
