+++ tests/download/viteScaffolder.test.mjs
@@ -2,7 +2,7 @@ import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
 import fs from 'fs';
 import path from 'path';
 import { fileURLToPath } from 'url';
-import { scaffoldViteProject } from '../src/viteScaffolder.js';
+import { scaffoldViteProject } from '../../src/download/viteScaffolder.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
